requirements.txt
discord.py
pyyaml
python-dotenv
Pillow